+++ src-tauri/src/lib.rs
@@ -12,6 +12,7 @@ pub fn run() {
         .plugin(tauri_plugin_dialog::init())
         .plugin(tauri_plugin_updater::Builder::new().build())
         .plugin(tauri_plugin_process::init())
+        .plugin(tauri_plugin_notification::init())
         .manage(PtyManager::default())
         .manage(AgentManager::default())
         .invoke_handler(tauri::generate_handler![